src/ingest_firehose/requirements.txt
boto3
orjson
pandas
svix-ksuid
fastparquet
zstandard